EDU-395 refactor: ChromaDB Client 분리 및 recommend를 recommandation으로 변경
